refactor: extract image download method
- Vision API 요청 데이터로 변환할 때 필요한 이미지 다운로드 메서드 분리
